requirements.txt
absl-py==0.15.0
asgiref==3.4.1
astor==0.8.0
astunparse==1.6.3
attrs==19.3.0
awsebcli==3.20.3
backcall==0.1.0
bleach==3.1.0
botocore==1.23.54
cached-property==1.5.2
cachetools==4.2.4
cement==2.8.2
certifi==2019.9.11
chardet==3.0.4
charset-normalizer==2.0.10
clang==5.0
Click==7.0
colorama==0.4.3
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.13
decorator==4.4.0
defusedxml==0.6.0
distlib==0.3.4
Django==3.2.13
django-ckeditor==6.3.2
django-js-asset==2.0.0
django-phonenumber-field==6.0.0
django-richtextfield==1.6.1
entrypoints==0.3
filelock==3.4.1
Flask==1.1.1
Flask-Cors==3.0.8
flatbuffers==1.12
future==0.16.0
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.1.0
idna==2.8
importlib-metadata==0.23
importlib-resources==5.4.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.10.0
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.1.1.26
opt-einsum==3.3.0
pathspec==0.9.0
phonenumbers==8.12.48
Pillow==6.2.0
pipenv==2022.4.8
platformdirs==2.4.0
protobuf==3.10.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.0.1
pypiwin32==223
python-dateutil==2.8.2
pytz==2022.1
pywin32==303
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
tornado==6.0.3
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.8
virtualenv==20.14.1
virtualenv-clone==0.5.7
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0
